sys_health_monitor/requirements.txt
psutil>=5.9.0
PyYAML>=6.0
rich>=13.0
